fix findService
