+++ src/main/java/com/gukbit/controller/UserController.java
@@ -132,97 +132,4 @@ public class UserController {
         return "false";
     }
 
-    @GetMapping("/findId")
-    public String findId() {
-        return "view/user/find-id";
-    }
-
-    @PostMapping("/findIdByTel")
-    public String findIdByTel(@RequestParam("tel") String tel, Model model) {
-        String message = userService.findIdByTel(tel);
-        model.addAttribute("message", message);
-        return "view/user/find-id-result";
-    }
-
-    @PostMapping("/findIdByEmail")
-    public String findIdByEmail(@RequestParam("email") String email, Model model) {
-        String message = userService.findIdByEmail(email);
-        model.addAttribute("message", message);
-        return "view/user/find-id-result";
-    }
-
-    @GetMapping("/findPw")
-    public String findPwAuth() {
-        return "view/user/find-pw";
-    }
-
-    @PostMapping("/findPwId")
-    public String findPwId(@RequestParam("id") String id, Model model) {
-        int count = 0;
-        if (id != null) {
-            count = userService.idCheck(id);
-        }
-
-        if (count != 0) {
-            model.addAttribute("userId", id);
-            return ("view/user/find-pw-auth");
-        } else {
-            model.addAttribute("message", "회원 정보를 찾을 수 없습니다.");
-            return ("view/user/find-pw-fail");
-        }
-    }
-
-    @PostMapping("/emailGetCode")
-    @ResponseBody
-    public String emailGetCode(@RequestParam("id") String id, @RequestParam("email") String email, Model model) {
-        if (userService.checkEmail(id, email) == 1) {
-            String code = mailService.sendEmailMessage(email);
-            model.addAttribute("code", code);
-            return code;
-        } else {
-            return "회원정보가 일치하지 않습니다.";
-        }
-    }
-
-    @PostMapping("/telGetCode")
-    @ResponseBody
-    public String telGetCode(@RequestParam("id") String id, @RequestParam("tel") String tel, Model model) {
-        if (userService.checkTel(id, tel) == 1) {
-            String code = mailService.sendTelMessage(tel);
-            model.addAttribute("code", code);
-            return code;
-        } else {
-            return "회원정보가 일치하지 않습니다.";
-        }
-    }
-
-    @PostMapping("/findPwEmail")
-    public String findPwEmail(@RequestParam("code") String code, Model model) {
-        if (mailService.getUserIdByEmailCode(code).equals("fail")) {
-            model.addAttribute("message", "인증코드를 다시 한 번 확인해주세요.");
-            return ("view/user/find-pw-fail");
-        } else {
-            model.addAttribute("userId", mailService.getUserIdByEmailCode(code));
-            return ("view/user/find-pw-success");
-        }
-    }
-
-    @PostMapping("/findPwTel")
-    public String findPwTel(@RequestParam("code") String code, Model model) {
-        if (mailService.getUserIdByTelCode(code).equals("fail")) {
-            model.addAttribute("message", "인증코드를 다시 한 번 확인해주세요.");
-            return ("view/user/find-pw-fail");
-        } else {
-            model.addAttribute("userId", mailService.getUserIdByTelCode(code));
-            return ("view/user/find-pw-success");
-        }
-    }
-
-    @PostMapping("/changePw")
-    public String changePw(@RequestParam("id") String id, @RequestParam("password") String password) {
-        userService.changePassword(id, password);
-        return "redirect:/";
-
-    }
-
 }
