+++ src/main/resources/templates/view/user/find-pw-success.html
@@ -19,7 +19,7 @@
         <div class="d-flex align-items-end justify-content-start" id="nav-find">
             <h1>비밀번호 재설정</h1>
         </div>
-        <form class="wrapped-findform" style="height: 200px;" method="post" th:action="@{/user/changePw}">
+        <form class="wrapped-findform" style="height: 200px;" method="post" th:action="@{/find/changePw}">
             <h5 style="margin-bottom: 20px;">비밀번호를 다시 설정해주세요</h5>
             <div class="d-flex justify-content-center align-items-center">
                 <i class="bi bi-lock"></i>
