+++ src/main/resources/templates/view/user/find-pw.html
@@ -21,7 +21,7 @@
         </div>
         <div class="wrapped-findform" style="height: 200px;">
             <h5>찾고자 하는 아이디를 입력해주세요</h5>
-            <form method="post" th:action="@{/user/findPwId}" class="d-flex justify-content-center align-items-center justify-content-start wrapped-find-input">
+            <form method="post" th:action="@{/find/findPwId}" class="d-flex justify-content-center align-items-center justify-content-start wrapped-find-input">
                 <i class="bi bi-person"></i>
                 <h1 class="d-flex justify-content-center align-items-center find-param">ID</h1><input type="text"
                 class="find-input" th:name="id"><button class="btn btn-primary find-button" type="submit">찾기</button>
